resume-analyzer/backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6
PyPDF2==3.0.1
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
email-validator==2.0.0
